+++ src/ArduinoJson/Data/Encoding.hpp
@@ -7,8 +7,6 @@
 
 #pragma once
 
-#include "../Print.hpp"
-
 namespace ArduinoJson {
 namespace Internals {