+++ test/Misc/StringBuilder.cpp
@@ -10,18 +10,10 @@
 
 using namespace ArduinoJson::Internals;
 
-TEST_CASE("StringBuilder") {
-  char output[20];
-  StaticStringBuilder sb(output, sizeof(output));
-
+template <typename StringBuilder, typename String>
+void common_tests(StringBuilder& sb, const String& output) {
   SECTION("InitialState") {
     REQUIRE(std::string("") == output);
-  }
-
-  SECTION("OverCapacity") {
-    REQUIRE(19 == sb.print("ABCDEFGHIJKLMNOPQRSTUVWXYZ"));
-    REQUIRE(0 == sb.print("ABC"));
-    REQUIRE(std::string("ABCDEFGHIJKLMNOPQRS") == output);
   }
 
   SECTION("EmptyString") {
@@ -39,4 +31,23 @@ TEST_CASE("StringBuilder") {
     REQUIRE(4 == sb.print("EFGH"));
     REQUIRE(std::string("ABCDEFGH") == output);
   }
+}
+
+TEST_CASE("StaticStringBuilder") {
+  char output[20];
+  StaticStringBuilder sb(output, sizeof(output));
+
+  common_tests(sb, static_cast<const char*>(output));
+
+  SECTION("OverCapacity") {
+    REQUIRE(19 == sb.print("ABCDEFGHIJKLMNOPQRSTUVWXYZ"));
+    REQUIRE(0 == sb.print("ABC"));
+    REQUIRE(std::string("ABCDEFGHIJKLMNOPQRS") == output);
+  }
+}
+
+TEST_CASE("DynamicStringBuilder") {
+  std::string output;
+  DynamicStringBuilder<std::string> sb(output);
+  common_tests(sb, output);
 }